src/utils/git_utils/requirements.txt
GitPython>=3.1.0
PyGithub>=2.1.0
loguru>=0.7.0
pytest>=7.0.0
pytest-cov>=4.1.0